+++ server/cmd/server/main.go
@@ -22,7 +22,7 @@ import (
 
 )
 
-const AppVersion = "0.4.4"
+const AppVersion = "0.4.5"
 
 func init() {
 	dbal.RegisterDialect(dbal.EnginePostgres, func() dbal.Dialect { return postgres.New() })